feat: searching restaurant (#61)
* feat: improve search functionality for restaurants

The code changes include adding a new endpoint `/restaurant/search` to the API that allows users to search for restaurants based on name and city. The search functionality filters the restaurants based on the provided query parameters and returns the matching results. This enhances the user experience by providing a more targeted and efficient way to find specific restaurants.

* test: added unit tests  for searching restaurant feature

* feat: Add endpoints for restaurant statistics

The code changes include adding new endpoints to the API for retrieving various statistics related to restaurants. These endpoints include `/restaurants/count` to get the total number of restaurants, `/restaurants/numberOfCities` to get the number of unique cities with restaurants, `/restaurants/cities` to get a list of all cities with restaurants, and `/restaurants/numberOfOpinions` to get the total number of opinions across all restaurants. These statistics provide valuable insights into the restaurant data and enhance the functionality of the application.