feat: add endpoint for scraping data
The code changes include adding a new endpoint `/scraping-data` to the API that allows users to retrieve scraped data from the Firebase Firestore. This endpoint fetches the data from the "scrapping-data" collection and returns it as a JSON response. This feature enhances the functionality of the application by providing users with access to the scraped data.