feat: optimizations last increment (#67)
* feat: add endpoint for updating restaurant count

The code changes include adding a new endpoint `/restaurant/:restaurantId/updateCount` to the API that allows users to update the count of a specific restaurant. This endpoint accepts the `restaurantId` as a parameter and increments the count by 1. This feature enhances the functionality of the application by providing users with the ability to track the number of times a restaurant has been visited.

Co-authored-by: Nicolás Sánchez Mendoza <nicsanmen@users.noreply.github.com>

* test: added unit tests to new endpoints

* feat: add endpoint for scraping data

The code changes include adding a new endpoint `/scraping-data` to the API that allows users to retrieve scraped data from the Firebase Firestore. This endpoint fetches the data from the "scrapping-data" collection and returns it as a JSON response. This feature enhances the functionality of the application by providing users with access to the scraped data.

* feat: sort feature matrix by ownerId in recommendations.js

This commit sorts the feature matrix in the recommendations.js file by ownerId in descending order. This change enhances the functionality of the application by ensuring that restaurants with ownerId appear first in the matrix.

---------

Co-authored-by: Nicolás Sánchez Mendoza <nicsanmen@users.noreply.github.com>